refactor: restructure server startup process and update package scripts
