+++ src/index.ts
@@ -1,0 +1,9 @@
+import { startServer } from './server';
+
+startServer()
+  .then(() => {
+    console.log('Server is running');
+  })
+  .catch(error => {
+    console.error('Failed to start server:', error);
+  });
