+++ src/server.ts
@@ -1,20 +1,39 @@
+import { Server } from 'node:http';
 import { createServer } from './config/websocket';
-import { setupProcessHandlers } from './utils/processHandlers';
+import { setupProcessHandlers } from './utils/server/processHandlers';
 import Logger from './logger';
 
-async function startServer(): Promise<void> {
+/**
+ * Starts the server and sets up necessary process handlers.
+ *
+ * This function performs the following steps:
+ * 1. Sets up process handlers for graceful shutdown and error handling.
+ * 2. Creates and starts the server using the `createServer` function.
+ * 3. Logs a message indicating that the server has started successfully.
+ *
+ * @returns {Promise<Server>} A promise that resolves to the started server instance.
+ *
+ * @example
+ * ```typescript
+ * startServer()
+ *   .then(server => {
+ *     console.log('Server is running');
+ *   })
+ *   .catch(error => {
+ *     console.error('Failed to start server:', error);
+ *   });
+ * ```
+ */
+async function startServer(): Promise<Server> {
   try {
     setupProcessHandlers();
-    await createServer();
+    const server = await createServer();
     Logger.info(`Server started successfully`);
+    return server;
   } catch (error) {
-    Logger.error('Fatal error during server startup:', error as Error);
-    process.exit(1);
+    Logger.error('Failed to start server:', error as Error);
+    throw error;
   }
 }
 
-// Start the server
-startServer().catch(error => {
-  Logger.error('Unexpected error:', error);
-  process.exit(1);
-});
+export { startServer };
